// src/app/openapi-gen/model/todoItem.ts
/**
 * API for Todo App
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TodoItem { 
    itemId?: number;
    listId?: string;
    taskName: string;
    createdAt?: string;
    done?: boolean;
}